refactor(indexer): move runtime tests out of src
