--- apps/indexer/scripts/placeholder.mjs
@@ -1,3 +1,0 @@
-console.log(
-  "The Datalens-native indexer runtime will be added in a follow-up issue."
-);
