+++ apps/indexer/src/runner.rs
@@ -457,7 +457,7 @@ where
     }
 }
 
-fn page_rows(rows: serde_json::Value) -> Result<Vec<serde_json::Value>, IndexerRunnerError> {
+pub fn page_rows(rows: serde_json::Value) -> Result<Vec<serde_json::Value>, IndexerRunnerError> {
     match rows {
         serde_json::Value::Array(rows) => Ok(rows),
         serde_json::Value::Object(mut object) => {
@@ -772,51 +772,5 @@ fn checkpoint(identity: IndexerCheckpointIdentity, start_block: i64) -> IndexerC
         last_error: None,
         lock_owner: None,
         locked_at: None,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use serde_json::json;
-
-    use super::*;
-
-    #[test]
-    fn test_page_rows_accepts_bare_array_response() {
-        let rows = page_rows(json!([{"block_number": 1}, {"block_number": 2}]))
-            .expect("bare rows are accepted");
-
-        assert_eq!(rows.len(), 2);
-    }
-
-    #[test]
-    fn test_page_rows_accepts_live_datalens_nested_response() {
-        let rows = page_rows(json!({
-            "dataset_key": {
-                "family": "Evm",
-                "name": "logs"
-            },
-            "rows": {
-                "dataset": "logs",
-                "rows": [
-                    {"block_number": 5873379}
-                ]
-            }
-        }))
-        .expect("live Datalens nested rows are accepted");
-
-        assert_eq!(rows, vec![json!({"block_number": 5873379})]);
-    }
-
-    #[test]
-    fn test_page_rows_rejects_malformed_response() {
-        let error = page_rows(json!({"rows": {"dataset": "logs"}}))
-            .expect_err("missing nested rows should fail");
-
-        assert!(
-            error
-                .to_string()
-                .contains("Datalens log query returned invalid rows payload")
-        );
     }
 }
